[MOD] SignUp.jsx -> component 분리, [ADD] SignUpComponents folder, 하위 styled-components folder, 내부 SignUpStyle.jsx, 외부 Contents.jsx, Sidebar.jsx
