+++ src/components/SignUpComponents/Sidebar.jsx
@@ -1,0 +1,6 @@
+import { AuthSidebar } from "./styled-components/SignUpStyle";
+
+function Sidebar() {
+  return <AuthSidebar />;
+}
+export default Sidebar;
